- damageDealt() method completed s - Created a few S1/S2 cases

+++ wyvernRunner/src/main/java/com/wyvernrunner/wicket/simulator/TempEffects/Bleed.java
@@ -3,7 +3,7 @@ package com.wyvernrunner.wicket.simulator.TempEffects;
 import com.wyvernrunner.wicket.simulator.Player;
 
 public class Bleed {
-    private final double type = 1;
+    private final int type = 1;
     private int duration;
 
     public Bleed(int dur) {
